requirements.txt
fastapi
uvicorn
pydantic-ai